+++ src/GafferScene/LocaliseAttributes.cpp
@@ -50,6 +50,7 @@ LocaliseAttributes::LocaliseAttributes( const std::string &name )
 {
 	storeIndexOfNextChild( g_firstPlugIndex );
 	addChild( new StringPlug( "attributes", Plug::In, "*" ) );
+	addChild( new BoolPlug( "includeGlobalAttributes", Plug::In, false ) );
 }
 
 LocaliseAttributes::~LocaliseAttributes()
@@ -66,18 +67,31 @@ const Gaffer::StringPlug *LocaliseAttributes::attributesPlug() const
 	return getChild<StringPlug>( g_firstPlugIndex );
 }
 
+Gaffer::BoolPlug *LocaliseAttributes::includeGlobalAttributesPlug()
+{
+	return getChild<BoolPlug>( g_firstPlugIndex + 1 );
+}
+
+const Gaffer::BoolPlug *LocaliseAttributes::includeGlobalAttributesPlug() const
+{
+	return getChild<BoolPlug>( g_firstPlugIndex + 1 );
+}
+
 bool LocaliseAttributes::affectsProcessedAttributes( const Gaffer::Plug *input ) const
 {
 	return
 		AttributeProcessor::affectsProcessedAttributes( input ) ||
-		input == attributesPlug()
+		input == attributesPlug() ||
+		input == includeGlobalAttributesPlug() ||
+		( input == inPlug()->globalsPlug() && !includeGlobalAttributesPlug()->isSetToDefault() )
 	;
 }
 
 void LocaliseAttributes::hashProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const
 {
 	AttributeProcessor::hashProcessedAttributes( path, context, h );
-	h.append( inPlug()->fullAttributesHash( path ) );
+	h.append( inPlug()->fullAttributesHash( path, /* withGlobalAttributes = */ includeGlobalAttributesPlug()->getValue() ) );
+	includeGlobalAttributesPlug()->hash( h );
 	attributesPlug()->hash( h );
 }
 
@@ -92,7 +106,7 @@ IECore::ConstCompoundObjectPtr LocaliseAttributes::computeProcessedAttributes( c
 	CompoundObjectPtr result = new CompoundObject;
 	result->members() = inputAttributes->members();
 
-	ConstCompoundObjectPtr fullAttributes = inPlug()->fullAttributes( path );
+	ConstCompoundObjectPtr fullAttributes = inPlug()->fullAttributes( path, /* withGlobalAttributes = */ includeGlobalAttributesPlug()->getValue() );
 	for( const auto &attribute : fullAttributes->members() )
 	{
 		if( StringAlgo::matchMultiple( attribute.first, attributes ) )